Move Rocket api code to a standalone crate
