+++ examples/rocket_example/README.md
@@ -2,9 +2,9 @@
 
 # Rocket with SeaORM example app
 
-1. Modify the `url` var in `Rocket.toml` to point to your chosen database
+1. Modify the `url` var in `api/Rocket.toml` to point to your chosen database
 
-1. Turn on the appropriate database feature for your chosen db in `Cargo.toml` (the `"sqlx-postgres",` line)
+1. Turn on the appropriate database feature for your chosen db in `core/Cargo.toml` (the `"sqlx-postgres",` line)
 
 1. Execute `cargo run` to start the server
