+++ examples/rocket_example/api/static/css/skeleton.css
