+++ examples/rocket_example/api/static/images/favicon.png
