+++ examples/rocket_example/api/templates/base.html.tera
