+++ examples/rocket_example/api/templates/error/404.html.tera
