+++ examples/rocket_example/api/templates/index.html.tera
