+++ examples/rocket_example/api/templates/new.html.tera
